+++ src/index.html
@@ -31,8 +31,8 @@
             <div class="all-comments"></div>
             <p>Add a comment</p>
             <form class="frm" id="addCmtForm">
-                <input type="text" placeholder="your name" class="name" id="name"/>
-                <textarea placeholder="your insights" class="text-area" id="text-area"></textarea>
+                <input type="text" placeholder="your name" class="name" id="name" required/>
+                <textarea placeholder="your insights" class="text-area" id="text-area" required></textarea>
                 <button type="submit" class="comment-btn">comment</button>
             </form>
         </div>
